type tweak
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@12071 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/applications/mod_http/mod_http.c
@@ -355,7 +355,7 @@ SWITCH_STANDARD_API(http_api_main)
 	json_response[m] = '\0';
 
 
-	switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_NOTICE, "RESERVED %ld BYTES, USED %ld BYTES, HTTP Response as JSON: %s\n", l, m, json_response);
+	switch_log_printf(SWITCH_CHANNEL_LOG, SWITCH_LOG_NOTICE, "RESERVED %u BYTES, USED %u BYTES, HTTP Response as JSON: %s\n", (unsigned)l, (unsigned) m, json_response);
 
 
 	stream->write_function(stream, "%s\n", json_response);
